phase-1to3: end-to-end network flow + cipher fix + UI polish
Brings the wire protocol up end-to-end against upstream Kinoko v95:

Network layer (MapleClaude.Net)
- Crypto: SendSentinel=GAME_VERSION (=95), RecvSentinel=0xFFFF-95,
  matches PacketEncoder.SEND_VERSION + PacketDecoder.RECV_VERSION exactly.
- Handshake IV ordering: first 4 wire bytes are client.sendIv
  (= server.recvIv per PacketChannelInitializer.connect(recvIv, sendIv)).
- Packet primitives: LE-short length-prefixed US-ASCII strings.
- Full InHeader/OutHeader enums + Packet/Session/Handlers/Crypto layout.
- ClientSession: System.IO.Pipelines socket I/O, locked per-direction IVs,
  thread-safe inbound queue drained on the MonoGame Update tick.
- MigrationCoordinator: login -> channel handoff via MigrateIn(20).
- LoginHandlers: CheckPasswordResult / WorldInformation / SelectWorldResult /
  CheckDuplicatedIDResult / CreateNewCharacterResult / DeleteCharacterResult /
  SelectCharacterResult / AliveReq -> typed events.
- FieldHandlers: SetField(141) + AliveReq, raises OnSetField with decoded
  CharacterStat + AvatarLook.
- AvatarCodec: byte-exact CharacterStat + AvatarLook round-trip.
- MovePathEncoder: full UserMove(44) move-path serializer (NORMAL / JUMP /
  TELEPORT / STAT_CHANGE / START_FALL_DOWN / FLYING_BLOCK / ACTION).

Stages (real network wired)
- LoginStage: ConnectAsync -> CheckPassword(1) -> CheckPasswordResult(0) ->
  WorldSelectStage. Yellow-on-dark status label so it's readable.
- WorldSelectStage: WorldInfoRequest(4) / SelectWorld(5) / SelectWorldResult(11).
- CharSelectStage: SelectCharacter(19) / SelectCharacterResult(12) ->
  MigrationCoordinator.BeginMigrateAsync.
- GameStage: subscribes to OnSetField, loads Map.wz scene via FieldScene,
  spawns PlayerController on a foothold, sends UserMove(44) every ~100 ms.
- PinStage scaffold (dormant; Kinoko sets bSkipPinCode = true).

Map + Character
- Foothold / Portal / MapInfo / TileInfo / FieldScene / Camera2D.
- CharacterRenderer + Stance + PlayerController (walk + jump + gravity +
  foothold ground-snap) + PlayerInput.

UI polish (LoginStage)
- TextField now supports caret + selection + Ctrl-A / Ctrl-C / Ctrl-V /
  Ctrl-X + Shift+arrows + Home/End/Delete + horizontal scroll so the caret
  stays visible past the field width.
- ClipboardHelper uses direct user32 P/Invoke so copy/paste works from
  MonoGame's MTA main thread (WinForms Clipboard requires STA and was
  silently throwing).
- IME toggle: Right Alt drives the Korean IME's Hangul-Yeong toggle via
  ImmGetDefaultIMEWnd + SendMessage(WM_IME_CONTROL, IMC_SETCONVERSIONMODE)
  since SDL2 swallows the raw Right Alt key.

Build / project layout
- SelfContained / PublishSingleFile gated behind IsAutoPublishing so the
  dev build stays RID-less (VS 2026 was looking for project-ref'd assemblies
  at win-x64-suffixed obj paths).
- IDE0005 / IDE0011 dropped to suggestion so external contributions with
  cosmetic unused-using noise don't fail TreatWarningsAsErrors.
- Privacy guard: zero hits on the local-only token grep.

Tests: 86/86 pass (cipher tests assert the wire-correct sentinel values;
new InPacket/OutPacket/MovePathEncoder/AvatarLook/HandshakeReader/
PacketRouter primitive tests added).

README + roadmap: Phase 1-3 marked shipped; Phase 3.5 added covering the
in-game cosmetic UI panels (StatusBar / Inventory / SkillBook / KeyConfig /
WorldMap / etc.) that landed via two contributor PRs and currently render
with placeholder data; Phases 4-10 list which packet sets each one needs.

+++ src/MapleClaude/UI/Game/StatusMessenger.cs
@@ -110,7 +110,7 @@ public sealed class StatusMessenger : GamePanel
 
             var sz  = _font.Measure(m.Text);
             var bgR = new Rectangle((int)Position.X - 2, (int)py - 1, (int)sz.X + 4, (int)sz.Y + 2);
-            sb.Draw(white, bgR, new Color(0, 0, 0, bgAlpha));
+            sb.Draw(white, bgR, new Color((byte)0, (byte)0, (byte)0, bgAlpha));
             _font.Draw(sb, m.Text, new Vector2(Position.X, py), textColor);
 
             py -= LineHeight;
